+++ src/main/java/com/prgms/allen/dining/domain/restaurant/RestaurantService.java
@@ -22,9 +22,6 @@ import com.prgms.allen.dining.domain.restaurant.dto.RestaurantSimpleRes;
 import com.prgms.allen.dining.domain.restaurant.entity.ClosingDay;
 import com.prgms.allen.dining.domain.restaurant.entity.Menu;
 import com.prgms.allen.dining.domain.restaurant.entity.Restaurant;
-import com.prgms.allen.dining.global.error.ErrorCode;
-import com.prgms.allen.dining.global.error.exception.NotFoundResourceException;
-import com.prgms.allen.dining.global.error.exception.RestaurantDuplicateCreationException;
 
 @Service
 @Transactional(readOnly = true)
